Majority of structural and editing work for #153 and #154
Still some discussions to have, and still some editing to do. (And MPD
work not entirely done!) But this edit essentially excises the AAT and
the RPT endpoint, adds a “saved consent token” design as an optional
substitute for gathering RqP consent in-band, and aligns the relevant
other artifacts as necessary.